requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.4
langchain-text-splitters>=0.2.0

faiss-cpu>=1.7.4
sentence-transformers>=2.6.0

transformers>=4.40.0
huggingface-hub>=0.23.0
tqdm>=4.66.0
numpy>=1.26.0

python-dotenv>=1.0.1
pypdf>=4.2.0

streamlit>=1.35.0